Split samples and real cloud support code
