+++ README.md
@@ -55,3 +55,17 @@ When built with `SIGNING_SUPPORT` a [dynamic engine](https://github.com/openssl/
 When built with `MBEDTLS_SUPPORT` a dynamic library is produced providing an API to let MbedTLS key use the signing key. See the [header file](include/TobMbedtls.h) for the details.
 
 In a resource-constrained application you most likely don't want to use this library, and probably have your own way to connect to the modem/SIM. In this way you can statically link to the low-level library. The [API](include/TbMbedtlsLL.h) allows you to substitute your own implementation of the [interface to the modem](external_libs/tob_sim/common/inc/SEInterface.h). The library doesn't use dynamic memory or multithreading.
+
+# Sample Azure IoT Registration Client Helper
+
+`azure_dps_registerer` is a simple client that uses the Azure IoT C SDK to register a device using the Device Provisioning Service.
+
+The client, when run, will read the DPS ID Scope either from a file (by default, `/home/pi/azure_id_scope.txt`) or an AZURE_ID_SCOPE environment variable if set.  It will use the Trust Onboard `available` certificate on your Twilio Trust Onboard SIM and register a device with Azure IoT Hub.  The output of this command will be a YAML encoded hash with connection information and the Trust Onboard `available` certificate and key in cleartext.
+
+The expected result is something similar to the following output:
+
+    status: SUCCESS
+    iothub_uri: ...
+    device_id: ...
+    certificate: "..."
+    key: "..."
