+++ samples/standalone/server/gencreds.sh
